# Cache the where clause extract_from_hash method