We know as soon as the associations are loaded if it s a column, reflection or aggregate, we we can check once and never again.
this way we only have to check what type a key is once. This avoid a reasonably large amount of work at runtime.

before:

```
-- create_table(:posts, {:force=>true})
   -> 0.0147s
-- create_table(:comments, {:force=>true})
   -> 0.0004s
Warming up --------------------------------------
Comment.where with 2 columns
                         1.631k i/100ms
Comment.where 1 association
                       614.000  i/100ms
Calculating -------------------------------------
Comment.where with 2 columns
                         16.215k (± 1.9%) i/s -     81.550k in   5.031218s
Comment.where 1 association
                          6.060k (± 2.1%) i/s -     30.700k in   5.068333s
```

after:

```
-- create_table(:posts, {:force=>true})
   -> 0.0150s
-- create_table(:comments, {:force=>true})
   -> 0.0004s
Warming up --------------------------------------
Comment.where with 2 columns
                         4.554k i/100ms
Comment.where 1 association
                       706.000  i/100ms
Calculating -------------------------------------
Comment.where with 2 columns
                         45.166k (± 2.3%) i/s -    227.700k in   5.044079s
Comment.where 1 association
                         6.964k (± 2.5%) i/s -     35.300k in   5.072544s
```

benchmark script:

<details>

```ruby
require "active_record"
require "logger"
require "stackprof"
require "benchmark/ips"

ActiveRecord::Base.establish_connection(adapter: "sqlite3", database: ":memory:")
ActiveRecord::Schema.define do
  create_table :posts, force: true do |t|
  end
  create_table :comments, force: true do |t|
    t.belongs_to :post
    t.string :title
    t.text :body
    t.integer :likes
    t.datetime :posted_at
  end
end

class Post < ActiveRecord::Base
  has_many :comments
end

class Comment < ActiveRecord::Base
  belongs_to :post
end

# puts "BEFORE WHERE"

# first case:
# Comment.where(post: { title: "hi" })
# # second case:
# puts "AFTER FIRST WHERE"

# Comment.where(post: Post.first(0))
# third case:
# Comment.group_by(:like_count).where(like_count: 10)
# fourth case:
# Comment.where(title: "best post ever!")

# binding.irb

Benchmark.ips do |x|
  x.report "Comment.where with 2 columns" do
    Comment.where(title: "best post ever!", post_id: 3)
  end

  x.report "Comment.where 1 association" do
    Comment.where(post: Post.first)
  end
end
```

</details