@HParker 2022-06-01
This performance difference is really trivial after removing the singularize.

Updated performance numbers:

before:

```
-- create_table(:posts, {:force=>true})
   -> 0.0073s
-- create_table(:comments, {:force=>true})
   -> 0.0002s
Warming up --------------------------------------
Comment.where with 2 columns
                         8.960k i/100ms
Comment.where 1 association
                         1.448k i/100ms
Calculating -------------------------------------
Comment.where with 2 columns
                         89.642k (± 0.2%) i/s -    456.960k in   5.097610s
Comment.where 1 association
                         14.501k (± 0.4%) i/s -     73.848k in   5.092721s
```

after:

```
-- create_table(:posts, {:force=>true})
   -> 0.0073s
-- create_table(:comments, {:force=>true})
   -> 0.0002s
Warming up --------------------------------------
Comment.where with 2 columns
                         9.186k i/100ms
Comment.where 1 association
                         1.436k i/100ms
Calculating -------------------------------------
Comment.where with 2 columns
                         90.688k (± 0.6%) i/s -    459.300k in   5.064802s
Comment.where 1 association
                         14.324k (± 0.5%) i/s -     71.800k in   5.012710s
```

I think we likely don't need this as long as we avoid the singularize call before checking if it is a column.